:sparkles: perf runner api (:exclamation: breaking change)
This is a HUGE commit, the original branch  was intendend to be a sandbox, but it eventually became something.

This commit introduces the ability to trace several runs in a file. A client runner orchestrates the execution and notifies a server runner. The server communicates with a driver that start/stop tracing and collect events.

The execution flow has changed for a file. Before files were traced in batch, then reports were created and displayed. For obvious reasons, this is not scalable. The new flow does tracing -> reporting for each file.

The global timeout has been removed for now. The user is in control of the execution time of each run. A timeout guard will be implemented later which will skip the next runs if one times out (and later again it will resume to next one).