:beetle: :art: fix duplicate event listeners, introduce director
trace-file role is to orchestrate the communication between the runner, driver and reporter. To do that it needs to subscribe to events and react to them. But it also need to store transitive data and unscribe to events.

It makes sense to replace the stateless reportFile function with a stateful Director class.